+++ include/lib/rule.hpp
@@ -1,0 +1,54 @@
+// MIT License
+//
+// Copyright (c) 2025
+//
+// Kavya Chopra (chopra.kavya04@gmail.com)
+// Cong Li (cong.li@inf.ethz.ch)
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+#ifndef REIFY_RULE_HPP
+#define REIFY_RULE_HPP
+
+#include "lib/lang.hpp"
+
+struct Rule {
+  Rule(std::string locPrefix = "local_created_by_rule_without_proper_locPrefix") : locPrefix(locPrefix) {}
+  virtual ~Rule() = default;
+  virtual bool match(const symir::Stmt *stmt) const = 0;
+  virtual std::vector<symir::BlockBuilder::StmtID> rewrite(
+    symir::FunctBuilder *funBd,
+    symir::BlockBuilder *blockBd,
+    const symir::Stmt *stmt
+  ) = 0;
+
+  /// get a new local that is not yet used in the block with label `blockLabel`
+  const symir::VarDef *getNewLocal(symir::FunctBuilder *funBd, std::string blockLabel) {
+    std::string locName = this->locPrefix + "_" + std::to_string(varCounterMap[blockLabel]++);
+    const symir::VarDef *loc = funBd->FindVar(locName);
+    if (loc == nullptr) loc = funBd->SymUnInitLocal(locName);
+    return loc;
+  }
+
+protected:
+  std::string locPrefix;
+  std::map<std::string, size_t> varCounterMap;
+};
+
+#endif // REIFY_RULE_HPP
