+++ lib/transformations.cpp
@@ -39,122 +39,6 @@
 
 using namespace patternmatch;
 
-namespace {
-  std::vector<symir::Coef *> copyAccess(symir::FunctBuilder *funBd, const symir::VarUse *use) {
-    std::vector<symir::Coef *> access;
-    access.reserve(use->GetAccess().size());
-    for (const auto &coef : use->GetAccess()) {
-      if (auto c = funBd->FindSymbol(coef->GetName()); c != nullptr) {
-        Assert(
-            typeid(*coef) == typeid(symir::Coef),
-            "Symbol \"%s\" is already defined and is not a coefficient", coef->GetName().c_str()
-        );
-        access.push_back(dynamic_cast<symir::Coef *>(coef));
-      } else {
-        Panic("coeff not found in provided function builder");
-      }
-    }
-    return access;
-  }
-
-  symir::BlockBuilder::CondID triviallyFalseCond(symir::FunctBuilder *funBd, symir::BlockBuilder *blockBd) {
-    return blockBd->SymCond(
-      symir::Cond::OP_EQZ,
-      blockBd->SymAddExpr({
-        blockBd->SymCstTerm(funBd->SymI32Const(1), nullptr)
-      })
-    );
-  }
-
-  symir::BlockBuilder::CondID triviallyTrueCond(symir::FunctBuilder *funBd, symir::BlockBuilder *blockBd) {
-    return blockBd->SymCond(
-      symir::Cond::OP_EQZ,
-      blockBd->SymAddExpr({
-        blockBd->SymCstTerm(funBd->SymI32Const(0), nullptr)
-      })
-    );
-  }
-
-  // match helpers:
-  bool matchSubExprInAnyStmt(const symir::Stmt *stmt, const Pattern<const symir::Expr *> &E) {
-    return patternmatch::match(
-      stmt,
-      m_AnyStmt(
-        m_WildCard<const symir::VarUse *>(),
-        m_WildCard<std::vector<const symir::VarUse *>>(),
-        E,
-        m_Cond(E),
-        m_Any(m_Cond(E)),
-        m_WildCard<const symir::ModExpr *>(),
-        m_Any(m_AssStmt(
-          m_WildCard<const symir::VarUse *>(),
-          E
-        )),
-        m_Any<std::vector<const symir::Stmt *>>(m_Any(m_AssStmt(
-          m_WildCard<const symir::VarUse *>(),
-          E
-        )))
-      )
-    );
-  }
-
-} // namespace
-
-template<typename Node>
-void StmtReplacer<Node>::Visit(const Node &n) { StmtCopier::Visit(n); }
-
-template<typename Node>
-symir::BlockBuilder::StmtID StmtReplacer<Node>::CopyStmtWithReplacement(
-  const symir::Stmt *s, 
-  std::function<bool(const Node *)> matchFunction,
-  std::function<ExprID(symir::FunctBuilder *, symir::BlockBuilder *, const Node &, void **)> replaceFunction,
-  double randThreshold,
-  size_t replaceMax
-) {
-  this->matchFunction = matchFunction;
-  this->replaceFunction = replaceFunction;
-  this->randThreshold = randThreshold;
-  this->randUniform = Random::Get().UniformReal();
-  this->hasReplaced = false;
-
-  s->Accept(*this);
-
-  if (!this->hasReplaced) {
-    // if by change (e.g. randTheshold) we have not replaced anything we run it again with a threshold of 1 to guarentee a replacement
-    popStmt();
-    this->randThreshold = 1;
-    s->Accept(*this);
-  }
-
-  Assert(this->hasReplaced, "StmtReplacer should only be called on stmt that are guaranteed to be able to be replaced");
-
-  return popStmt();
-}
-
-template<> void StmtReplacer<symir::Expr>::Visit(const symir::Expr &e) {
-  if (this->match(e)) {
-    pushExpr(this->replace(e));
-  } else {
-    StmtCopier::Visit(e);
-  }
-}
-
-template<> void StmtReplacer<symir::Term>::Visit(const symir::Term &t) {
-  if (this->match(t)) {
-    pushTerm(this->replace(t));
-  } else {
-    StmtCopier::Visit(t);
-  }
-}
-
-template<> void StmtReplacer<symir::Cond>::Visit(const symir::Cond &c) {
-  if (this->match(c)) {
-    pushCond(this->replace(c));
-  } else {
-    StmtCopier::Visit(c);
-  }
-}
-
 void RewriteEngine::addRule(std::unique_ptr<Rule> rule, int weight) {
   rules.push_back(std::move(rule));
   weights.push_back(weight);
@@ -355,303 +239,6 @@ std::optional<Rule *> RewriteEngine::getRandomMatchingRule(const symir::Stmt *st
   int index = 0;
   while (selWeight > this->weights[matchingRules[index]]) selWeight -= this->weights[matchingRules[index++]];
   return this->rules[matchingRules[index]].get();
-}
-
-bool ConstProba::match(const symir::Stmt *stmt) const {
-  return matchSubExprInAnyStmt(stmt, m_Expr(m_Any(m_CstTerm(m_Solved(), m_NoVar()))));
-}
-
-std::vector<symir::BlockBuilder::StmtID> ConstProba::rewrite(
-  symir::FunctBuilder *funBd,
-  symir::BlockBuilder *blockBd,
-  const symir::Stmt *stmt
-) {
-  Log::Get().Out() << "Running ConstProba" << std::endl;
-
-  StmtReplacer rep = StmtReplacer<symir::Term>(funBd, blockBd);
-  const symir::VarDef *var = this->getNewLocal(funBd, blockBd->GetLabel());
-  std::function<symir::BlockBuilder::TermID(symir::FunctBuilder * ,symir::BlockBuilder *, const symir::Term &, void **)>
-    varInsertFun =
-      [&](symir::FunctBuilder *thisFunBd, symir::BlockBuilder *thisBlockBd, const symir::Term &t, void **data) {
-        *data = t.GetCoef();
-        return thisBlockBd->SymMulTerm(thisFunBd->SymI32Const(1), var);
-      };
-
-  symir::BlockBuilder::StmtID newStmt = rep.CopyStmtWithReplacement(
-    stmt,
-    make_matcher(const symir::Term *, m_CstTerm(m_Solved(), m_NoVar())),
-    varInsertFun,
-    0.25
-  );
-  symir::Coef *replacedCoef = static_cast<symir::Coef *>(rep.getExtractedDataRef());
-  Assert(replacedCoef != nullptr, "replacedCoef should never be nullptr");
-  Assert(replacedCoef->IsSolved(), "replacedCoef should never be unsolved");
-
-  Log::Get().Out() << "Replacing Const " << replacedCoef->GetI32Value() << " with " << var->GetName() << std::endl;
-
-  symir::BlockBuilder::StmtID assignStmts = blockBd->SymAssStmt(
-    var,
-    blockBd->SymExpr(
-      symir::Expr::OP_ADD,
-      { blockBd->SymCstTerm(replacedCoef, nullptr) }
-    )
-  );
-
-  return {assignStmts, newStmt};
-}
-
-
-bool AdditionFromConst::match(const symir::Stmt *stmt) const {
-  return matchSubExprInAnyStmt(stmt, m_Expr(m_Any(m_CstTerm(m_Solved(), m_NoVar()))));
-}
-
-std::vector<symir::BlockBuilder::StmtID> AdditionFromConst::rewrite(
-  symir::FunctBuilder *funBd,
-  symir::BlockBuilder *blockBd,
-  const symir::Stmt *stmt
-) {
-  Log::Get().Out() << "Running AdditionFromConst" << std::endl;
-
-  StmtReplacer rep = StmtReplacer<symir::Expr>(funBd, blockBd);
-  std::function<symir::BlockBuilder::ExprID(symir::FunctBuilder * ,symir::BlockBuilder *, const symir::Expr &, void **)>
-    varInsertFun =
-      [&](symir::FunctBuilder *thisFunBd, symir::BlockBuilder *thisBlockBd, const symir::Expr &e, void **data) {
-        std::vector<symir::BlockBuilder::TermID> termIds;
-        termIds.reserve(e.GetTerms().size() + 1);
-        bool hasReplaced = false;
-        for (auto term : e.GetTerms()) {
-          if (!hasReplaced && term->GetOp() == symir::Term::OP_CST) {
-            hasReplaced = true;
-            *data = term->GetCoef();
-            int target = term->GetCoef()->GetI32Value();
-            // v1 must be choosen s.t. |v1| < |target| and sign(v1) == sign(target)
-            // this ensures a UB free transformation since:
-            // if Sk is the prefix sum up to the target Term then
-            // |Sk op v1| < |Sk op target| and since Sk op target does not overflow neither does |Sk op v1|
-            int v1, v2;
-            if (target >= 0) {
-              v1 = Random::Get().Uniform(0, target)();
-            } else {
-              v1 = Random::Get().Uniform(target, -1)();
-            }
-            v2 = target - v1;
-            if (e.GetOp() == symir::Expr::OP_SUB) v2 = -v2;
-
-            termIds.push_back(blockBd->SymTerm(
-              symir::Term::OP_CST,
-              funBd->SymI32Const(v1),
-              nullptr, {})
-            );
-            termIds.push_back(blockBd->SymTerm(
-              symir::Term::OP_CST,
-              funBd->SymI32Const(v2),
-              nullptr, {})
-            );
-          } else {
-            termIds.push_back(symir::StmtCopier(funBd, blockBd).CopyTerm(term));
-          }
-        }
-
-      return blockBd->SymExpr(e.GetOp(), termIds);
-    };
-  symir::BlockBuilder::StmtID newStmt = rep.CopyStmtWithReplacement(
-    stmt,
-    make_matcher(const symir::Expr *, m_Expr(m_Any(m_CstTerm(m_Solved(), m_NoVar())))),
-    varInsertFun,
-    0.25
-  );
-  symir::Coef *replacedCoef = static_cast<symir::Coef *>(rep.getExtractedDataRef());
-  Assert(replacedCoef != nullptr, "replacedCoef should never be nullptr");
-  Assert(replacedCoef->IsSolved(), "replacedCoef should never be unsolved");
-
-  Log::Get().Out() << "Replacing Const " << replacedCoef->GetI32Value() << " with addition" << std::endl;
-
-  return {newStmt};
-}
-
-bool ForSumFromConst::match(const symir::Stmt *stmt) const {
-  return patternmatch::match(
-      stmt,
-      m_AssStmt(
-        m_WildCard<const symir::VarUse *>(),
-        m_Expr(m_One(m_CstTerm(m_Solved(), m_NoVar())))
-      )
-    );
-}
-
-std::vector<symir::BlockBuilder::StmtID> ForSumFromConst::rewrite(
-  symir::FunctBuilder *funBd,
-  symir::BlockBuilder *blockBd,
-  const symir::Stmt *stmt
-) {
-  Log::Get().Out() << "Running ConstToForSum" << std::endl;
-
-  const symir::AssStmt *assStmt = static_cast<const symir::AssStmt *>(stmt);
-  const auto use = assStmt->GetVar();
-  const auto def = use->GetDef();
-  const auto expr = assStmt->GetExpr();
-
-  Assert(
-    expr->GetTerms().size() == 1 && expr->GetTerm(0)->GetOp() == symir::Term::OP_CST,
-    "match function for ConstToForSum failed to uphold its promisses"
-  );
-
-  int32_t val = expr->GetTerm(0)->GetCoef()->GetI32Value();
-  int32_t loopCount, rest, randVal;
-  if (val > 0) {
-    randVal = Random::Get().Uniform(1, std::max(val / 2, 1))();
-    rest = val % randVal;
-    loopCount = val / randVal;
-  } else if (val == 0) {
-    randVal = 0;
-    rest = 0;
-    loopCount = 0;
-  } else {
-    randVal = Random::Get().Uniform(std::min(val / 2, -1), -1)();
-    rest = val % randVal;
-    loopCount = val / randVal;
-  }
-  Assert(loopCount >= 0, "loopCount must be larger then 0");
-  Log::Get().Out() << "Target: " << val << ", LoopCount: " << loopCount 
-                   << ", Additive Value: " << randVal << ", Remainder: " << rest << std::endl;
-
-  auto access = copyAccess(funBd, use);
-
-  symir::BlockBuilder::StmtID initAss = blockBd->SymAssStmt(
-    def,
-    blockBd->SymAddExpr({
-      blockBd->SymCstTerm(
-        funBd->SymI32Const(rest),
-        nullptr, {} 
-      )
-    }),
-    access
-  );
-
-  symir::BlockBuilder::StmtID addAss = blockBd->SymAssStmt(
-    def,
-    blockBd->SymAddExpr({
-      blockBd->SymAddTerm(
-        funBd->SymI32Const(randVal),
-        def, access
-      )
-    }),
-    access
-  );
-
-  auto loopVar = this->getNewLocal(funBd, blockBd->GetLabel());
-  symir::BlockBuilder::StmtID forSum = blockBd->SymForStmt(
-    // loop variable i
-    loopVar,
-    // loop condition (i - loopCount < 0)
-    blockBd->SymCond(symir::Cond::OP_LTZ, 
-      blockBd->SymAddExpr({
-        blockBd->SymMulTerm(
-          funBd->SymI32Const(1),
-          loopVar, {}
-        ),
-        blockBd->SymCstTerm(
-          funBd->SymI32Const(-loopCount),
-          nullptr, {}
-        )
-      })
-    ),
-    // int i = 0
-    blockBd->SymAddExpr({
-      blockBd->SymCstTerm(
-        funBd->SymI32Const(0),
-        nullptr
-      )
-    }),
-    // i += 1
-    blockBd->SymAddExpr({
-      blockBd->SymCstTerm(
-        funBd->SymI32Const(1),
-        nullptr
-      )
-    }),
-    {addAss}
-  );
-
-  return { initAss, forSum };
-}
-
-bool DeadCodeFromAssign::match(const symir::Stmt *stmt) const {
-  return patternmatch::match(stmt, m_AssStmt(m_WildCard<const symir::VarUse *>(), m_WildCard<const symir::Expr *>()));
-}
-
-std::vector<symir::BlockBuilder::StmtID> DeadCodeFromAssign::rewrite(
-  symir::FunctBuilder *funBd,
-  symir::BlockBuilder *blockBd,
-  const symir::Stmt *stmt
-) {
-  Log::Get().Out() << "Running AssToDeadCode" << std::endl;
-
-  const symir::AssStmt *assStmt = static_cast<const symir::AssStmt *>(stmt);
-  const auto use = assStmt->GetVar();
-  const auto def = use->GetDef();
-  const auto expr = assStmt->GetExpr();
-  auto access = copyAccess(funBd, use);
-
-  symir::BlockBuilder::ExprID exprId = symir::StmtCopier(funBd, blockBd).CopyExpr(expr);
-
-  int nrBranches = Random::Get().Uniform(this->minBranches, this->maxBranches)();
-  int trueBranch = Random::Get().Uniform(0, nrBranches - 1)();
-
-  Log::Get().Out() << "Building " << nrBranches << " branches, True branch: " << trueBranch << std::endl;
-
-  std::vector<symir::BlockBuilder::CondID> conds;
-  std::vector<std::vector<symir::BlockBuilder::StmtID>> sids;
-  conds.reserve(nrBranches - 1);
-  sids.reserve(nrBranches);
-  for (int i = 0; i < nrBranches; i++) {
-    if (i < nrBranches - 1) {
-      // not the else branch hence we create a condition
-      if (i == trueBranch) {
-        conds.push_back(triviallyTrueCond(funBd, blockBd));
-      } else {
-        conds.push_back(triviallyFalseCond(funBd, blockBd));
-      }
-    }
-    if (i == trueBranch) {
-      sids.push_back({ blockBd->SymAssStmt(def, exprId, access) });
-    } else {
-      // TODO: genrate more complex expression and take the allowUB member into account
-      sids.push_back({
-        blockBd->SymAssStmt(
-          def,
-          blockBd->SymAddExpr({
-            blockBd->SymCstTerm(
-              funBd->SymI32Const(Random::Get().Uniform(INT_MIN, INT_MAX)()),
-              nullptr
-            )
-          }),
-          access
-        ) 
-      });
-    }
-    
-  }
-
-  return { blockBd->SymIfStmt(conds, sids) };
-}
-
-bool VectorizerDeadAssignFromCopy::match(const symir::Stmt *stmt) const {
-  return patternmatch::match(stmt, m_AssStmt(m_WildCard<const symir::VarUse*>(), m_WildCard<const symir::Expr *>()));
-}
-
-std::vector<symir::BlockBuilder::StmtID> VectorizerDeadAssignFromCopy::rewrite(
-    symir::FunctBuilder *funBd,
-    symir::BlockBuilder *blockBd,
-    const symir::Stmt *stmt
-  ) {
-  auto copier = symir::StmtCopier(funBd, blockBd);
-  symir::BlockBuilder::StmtID origStmt = copier.CopyStmt(stmt);
-  symir::BlockBuilder::StmtID deadStmt = blockBd->SymAssStmt(
-    this->getNewLocal(funBd, blockBd->GetLabel()),
-    copier.CopyExpr(static_cast<const symir::AssStmt *>(stmt)->GetExpr())
-  );
-  return {deadStmt, origStmt};
 }
 
 std::vector<const symir::Term *> ConstQuery::query() {
