+++ src/revopt-fgen.cpp
@@ -177,12 +177,7 @@ public:
       )
     );
 
-    auto revopt = RewriteEngine();
-    revopt.addRule(std::make_unique<ConstProba>(), 3);
-    revopt.addRule(std::make_unique<AdditionFromConst>(), 3);
-    revopt.addRule(std::make_unique<ForSumFromConst>(), 2);
-    revopt.addRule(std::make_unique<DeadCodeFromAssign>(), 1);
-    revopt.addRule(std::make_unique<VectorizerDeadAssignFromCopy>(), 1);
+    auto revopt = RewriteEngine::Default();
     revopt.run(builder.get(), bblBd, 20);
 
     auto cstTerms = ConstQuery(builder.get(), bblBd).query();
